feat(server): move http related functions in the another module
